move LogWithMetadata to core
